Implement iframe server that makes subscription request
When the frame is ready to receive subscription requests, it posts a message of
type `ready` to its parent.

On receipt of a `subscribe` message, it POSTs to
`/api/v2/widget/<widgetId>/subscribe/<mobileNumber>`, where `widgetId` and
`mobileNumber` are obtained from the message. When it gets a response from the
server, it posts a message of type `result` to its parent, with a `success`
property that is true on success and false otherwise.